+++ src/client/router/RoutesConfig.js
@@ -1,18 +1,26 @@
-import App from '../views/_navigation/index.js';
+import App from './App.js';
+//
 import Home from '../views/_pages/Home/index.js';
 import Posts from '../views/_pages/Posts/index.js';
 import PostDetail from '../views/_pages/Posts/PostDetail.js';
 import Todos from '../views/_pages/Todos/index.js';
+import NoMatch from '../views/_pages/404/index.js';
+import NestedRoutes from '../views/_pages/NestedRoutes/index.js';
 
 
-export default [
+const routesConfig = [
   {
     component: App,
     routes: [
       {
         path: "/",
         component: Home,
-        exact: true,
+        exact: true
+      },
+		
+      {
+        path: "/index",
+        component: Home
       },
 		
 	/* 
@@ -23,16 +31,31 @@ export default [
 		
       {
         path: "/posts",
-        component: Posts
+        component: Posts,
+		exact: true
       },
       {
         path: '/todos',
         component: Todos
       },
       {
-        path: '/post-detail/:post_id',
+        path: '/posts/:post_id',
         component: PostDetail
-      }     
+      },
+		
+	  /* In order to make nested routes valid, do not add the exact attribute here */
+	  {
+        path: '/nested-routes',
+        component: NestedRoutes
+      },	
+      {
+        path: '*',
+        component: NoMatch
+      }  	
+		
     ]
   }
 ];
+
+
+export default routesConfig;