+++ src/client/router/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { renderRoutes } from 'react-router-config';
-import Routes from './routes.js';
+import customRoutesConfig from './RoutesConfig.js';
 
 export default () => {
   return (
     <BrowserRouter>
-      <div>{renderRoutes(Routes)}</div>
+      <div id="main">{renderRoutes(customRoutesConfig)}</div>
     </BrowserRouter>
   );
 };